all files updated

+++ employeems/src/components/GetEmployees.jsx
@@ -1,5 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
+import NavBarComp from "./NavBarComp";
+import FooterComp from "./FooterComp";
 
 const cssStyle = {
   border: "1px solid black",
@@ -7,7 +9,7 @@ const cssStyle = {
 
 const GetEmployees = () => {
   let [employees, setEmployees] = useState([]);
-  const navigate=useNavigate();
+  const navigate = useNavigate();
 
   const getEmployees = async () => {
     let fetchData = await fetch("http://localhost:8080/api/getemployees");
@@ -22,32 +24,35 @@ const GetEmployees = () => {
 
   const deleteFunction = async (eid) => {
     try {
-      let deleteData = await fetch(
-        `http://localhost:8080/api/deleteemployee/${eid}`,
-        {
+      if (window.confirm("Kharena?") === true) {
+        await fetch(`http://localhost:8080/api/deleteemployee/${eid}`, {
           method: "DELETE", // Specify the DELETE method
-        }
-      );
-      getEmployees();
+        });
+
+        getEmployees();
+      }
     } catch (error) {
       console.log(error);
     }
   };
 
-  const editFunction =(eid)=>{
-    navigate(`/updateemployee/${eid}`)
-
-  }
+  const editFunction = (eid) => {
+    navigate(`/updateemployee/${eid}`);
+  };
 
   return (
     <>
-      <table style={cssStyle}>
-        <tr>
-          <th style={cssStyle}>Employee Id</th>
-          <th style={cssStyle}>Employee Name</th>
-          <th style={cssStyle}>Employee Age</th>
-          <th style={cssStyle}>Employee Salary</th>
-        </tr>
+      <NavBarComp />
+      <table className="table">
+        <thead className="thead-dark">
+          <tr>
+            <th style={cssStyle}>Employee Id</th>
+            <th style={cssStyle}>Employee Name</th>
+            <th style={cssStyle}>Employee Age</th>
+            <th style={cssStyle}>Employee Salary</th>
+          </tr>
+        </thead>
+
         {employees.map((currentEmployee) => {
           return (
             <>
@@ -56,19 +61,26 @@ const GetEmployees = () => {
                 <td style={cssStyle}>{currentEmployee.ename}</td>
                 <td style={cssStyle}>{currentEmployee.eage}</td>
                 <td style={cssStyle}>{currentEmployee.esalary}</td>
-                <button onClick={() => deleteFunction(currentEmployee.eid)}>
+                <button style={{backgroundColor: "red",color:"white", padding:"2px"}} onClick={() => deleteFunction(currentEmployee.eid)}>
                   Delete
                 </button>
-                <button onClick={() =>editFunction(currentEmployee.eid)}>Edit</button>
+                <button style={{backgroundColor: "#4caf50", padding:"2px"}} onClick={() => editFunction(currentEmployee.eid)}>
+                  Edit
+                </button>
               </tr>
             </>
           );
         })}
       </table>
 
-      <button onClick={()=>{
-        navigate("/")
-      }}>Home Page</button>
+      <button style={{backgroundColor:"blue",color:'white'}}
+        onClick={() => {
+          navigate("/home");
+        }}
+      >
+        Home Page
+      </button>
+      <FooterComp />
     </>
   );
 };
